[WIP] Qualitative ramps selector refactor
