+++ lib/assets/core/javascripts/cartodb3/components/form-components/editors/fill/input-color/input-color-value-content-view.js
@@ -3,9 +3,11 @@ var Backbone = require('backbone');
 var CoreView = require('backbone/core-view');
 var StackLayoutView = require('../../../../../components/stack-layout/stack-layout-view');
 var ColumnListView = require('../column-list-view');
-var quantificationMethodItemTemplate = require('./quantification-method-item.tpl');
-var InputColorRamps = require('./input-quantitative-ramps/input-color-ramps');
-var InputColorCategories = require('./input-categories/input-color-categories');
+var QuantificationMethodItemTemplate = require('./quantification-method-item.tpl');
+
+var InputQuantitativeRamps = require('./input-quantitative-ramps/main-view');
+var InputQualitativeRamps = require('./input-qualitative-ramps/main-view');
+
 var DefaultFillSettings = require('../default-fill-settings.json');
 var checkAndBuildOpts = require('../../../../../helpers/required-opts');
 var MetricsTracker = require('../../../../../components/metrics/metrics-tracker');
@@ -60,10 +62,10 @@ module.exports = CoreView.extend({
         createStackView: this._createColumnListView.bind(this)
       },
       {
-        createStackView: this._createInputColorRamps.bind(this)
+        createStackView: this._createInputQuantitativeRamps.bind(this)
       },
       {
-        createStackView: this._createInputColorCategories.bind(this)
+        createStackView: this._createInputQualitativeRamps.bind(this)
       }
     ]);
 
@@ -91,8 +93,8 @@ module.exports = CoreView.extend({
     this.addView(this._stackLayoutView);
   },
 
-  _createInputColorCategories: function (stackLayoutModel, opts) {
-    var view = new InputColorCategories({
+  _createInputQualitativeRamps: function (stackLayoutModel, opts) {
+    var view = new InputQualitativeRamps({
       model: this.model,
       query: this._query,
       columns: this._columns,
@@ -118,7 +120,7 @@ module.exports = CoreView.extend({
     var view = new ColumnListView({
       headerTitle: _t('form-components.editors.fill.quantification.title'),
       stackLayoutModel: stackLayoutModel,
-      itemTemplate: quantificationMethodItemTemplate,
+      itemTemplate: QuantificationMethodItemTemplate,
       columns: this._settings.quantifications.items,
       showSearch: false
     });
@@ -135,8 +137,8 @@ module.exports = CoreView.extend({
     return view;
   },
 
-  _createInputColorRamps: function (stackLayoutModel, opts) {
-    var view = new InputColorRamps({
+  _createInputQuantitativeRamps: function (stackLayoutModel, opts) {
+    var view = new InputQuantitativeRamps({
       hideTabs: this.options.hideTabs,
       model: this.model
     });
