+++ lib/assets/core/javascripts/cartodb3/components/form-components/editors/fill/input-color/input-qualitative-ramps/input-categories-list/input-color-categories-list-view.tpl
