+++ lib/assets/core/javascripts/cartodb3/components/form-components/editors/fill/input-color/input-qualitative-ramps/input-categories-ramps/input-categories-ramps-list-item-view.js
@@ -1,16 +1,14 @@
-var CoreView = require('backbone/core-view');
 var _ = require('underscore');
+var CoreView = require('backbone/core-view');
 
 module.exports = CoreView.extend({
-
   className: 'CDB-ListDecoration-item CustomList-item CustomList-item--invert js-listItem',
   tagName: 'li',
 
   events: {
     'mouseenter': '_onMouseEnter',
     'mouseleave': '_onMouseLeave',
-    'click .js-listItemLink': '_onClick',
-    'click .js-invert': '_onClickInvert'
+    'click .js-listItemLink': '_onClick'
   },
 
   initialize: function (opts) {
@@ -27,7 +25,6 @@ module.exports = CoreView.extend({
         _.extend({
           typeLabel: this.options.typeLabel,
           isSelected: this.model.get('selected'),
-          isInverted: this.model.get('inverted'),
           isDisabled: this.model.get('disabled'),
           isDestructive: this.model.get('destructive'),
           name: this.model.getName(),
@@ -40,8 +37,7 @@ module.exports = CoreView.extend({
     this.$el
       .attr('data-val', this.model.getValue())
       .toggleClass('is-disabled', !!this.model.get('disabled'))
-      .toggleClass('is-selected', !!this.model.get('selected'))
-      .toggleClass('is-inverted', !!this.model.get('inverted'));
+      .toggleClass('is-selected', !!this.model.get('selected'));
 
     return this;
   },
@@ -62,11 +58,5 @@ module.exports = CoreView.extend({
     } else {
       this.model.set('selected', true);
     }
-  },
-
-  _onClickInvert: function (e) {
-    this.killEvent(e);
-    this.model.set('inverted', !this.model.get('inverted'));
-    this.trigger('customEvent', 'invert', this.model, this);
   }
 });
