+++ lib/assets/core/javascripts/cartodb3/components/form-components/editors/fill/input-color/input-qualitative-ramps/input-categories-ramps/input-categories-ramps-list-item.tpl
@@ -1,6 +1,6 @@
 <div class="ColorBarWrapper--withInvertLink CDB-ListDecoration-itemLink CDB-ListDecoration-itemLink--double js-listItemLink <% if (isSelected) { %> is-selected <% } %>">
   <ul class="ColorBarContainer">
-    <% _.each(colors, function (color) { %>
+    <% _.each(name, function (color) { %>
     <li class="ColorBar ColorBar--spaceless" style="background-color: <%- color %>;"></li>
     <% }); %>
   </ul>
