+++ lib/assets/core/javascripts/cartodb3/components/form-components/editors/fill/input-color/input-qualitative-ramps/main-view.js
@@ -1,16 +1,17 @@
 var _ = require('underscore');
-var CDB = require('cartodb.js');
+var cdb = require('cartodb.js');
 var Backbone = require('backbone');
 var CoreView = require('backbone/core-view');
-var mainViewTemplate = require('./main-view.tpl');
-var CategoriesListView = require('./input-color-categories-list-view/input-color-categories-list-view');
-var InputCategoriesRamp = require('./input-categories-ramps/input-ramp-content-view');
+var MainViewTemplate = require('./main-view.tpl');
+
 var StackLayoutView = require('../../../../../../components/stack-layout/stack-layout-view');
+var AssetPickerView = require('../assets-picker/input-asset-picker-view');
+
+var CategoriesRampsView = require('./input-categories-ramps/input-categories-ramps-list-view');
+var CategoriesListView = require('./input-categories-list/input-color-categories-list-view');
 var InputColorPickerView = require('../input-color-picker/input-color-picker-view');
 
 var CartoColor = require('cartocolor');
-var AssetPickerView = require('../assets-picker/input-asset-picker-view');
-
 var MAX_VALUES = 10;
 var DEFAULT_COLORS = _.clone(CartoColor.Prism[MAX_VALUES + 1]); // max values + "others" color
 
@@ -26,22 +27,16 @@ var REQUIRED_OPTS = [
 ];
 
 var STEPS = {
-  RAMPS_PICKER_VIEW: 0,
-  RANGE_LIST_VIEW: 1,
+  CATEGORIES_RAMPS_VIEW: 0,
+  CATEGORIES_LIST_VIEW: 1,
   COLOR_PICKER_VIEW: 2,
-  IMAGE_PICKER_VIEW: 3 // TODO needed?
+  IMAGE_PICKER_VIEW: 3
 };
-
-function _quote (c) {
-  if (c && c !== true) {
-    return '"' + c.toString().replace(/"/g, '\\"').replace(/\n/g, '\\n') + '"';
-  }
-  return c;
-}
 
 module.exports = CoreView.extend({
   events: {
-    'click .js-back': '_onClickBack'
+    'click .js-back': '_onClickBack',
+    'click .js-quantification': '_onClickQuantification'
   },
 
   initialize: function (opts) {
@@ -55,18 +50,31 @@ module.exports = CoreView.extend({
     });
 
     this.listenTo(this._viewModel, 'change:status', this.render);
+
+    // If it comes from a quantification change
+    if (this.model.hasChanged('quantification') && this.model.get('quantification') === 'category') {
+      this._fetchColumns();
+    }
   },
 
   render: function () {
     this.clearSubViews();
     this.$el.empty();
 
-    this.$el.append(mainViewTemplate({
+    var column = this._getColumn();
+    var columnType = column && column.type;
+    var hideQuantification = !!(this._hideTabs && _.contains(this._hideTabs, 'quantification'));
+
+    this.$el.append(MainViewTemplate({
       status: this._viewModel.get('status'),
-      attribute: this.model.get('attribute')
+      columnType: columnType,
+      attribute: this.model.get('attribute'),
+      quantification: this.model.get('quantification') || 'category',
+      hideQuantification: hideQuantification
     }));
 
     this._generateStackLayoutView();
+
     return this;
   },
 
@@ -74,11 +82,11 @@ module.exports = CoreView.extend({
     var currentStep = this._viewModel.get('step');
     var stackViewCollection = new Backbone.Collection([
       {
-        selected: currentStep === STEPS.RAMPS_PICKER_VIEW,
-        createStackView: this._createCategoriesRampView.bind(this)
+        selected: currentStep === STEPS.CATEGORIES_RAMPS_VIEW,
+        createStackView: this._createCategoriesRampsView.bind(this)
       },
       {
-        selected: currentStep === STEPS.RANGE_LIST_VIEW,
+        selected: currentStep === STEPS.CATEGORIES_LIST_VIEW,
         createStackView: this._createRangeListView.bind(this)
       },
       {
@@ -123,7 +131,7 @@ module.exports = CoreView.extend({
     if (this._query) {
       this._setViewStatus('loading');
 
-      var sql = new CDB.SQL({
+      var sql = new cdb.SQL({
         user: this._configModel.get('user_name'),
         sql_api_template: this._configModel.get('sql_api_template'),
         api_key: this._configModel.get('api_key')
@@ -276,19 +284,19 @@ module.exports = CoreView.extend({
     return view;
   },
 
-  _createCategoriesRampView: function (stackLayoutModel, opts) {
-    var view = new InputCategoriesRamp({
+  _createCategoriesRampsView: function (stackLayoutModel, opts) {
+    var view = new CategoriesRampsView({
       maxValues: MAX_VALUES,
       model: this.model
     });
 
-    view.on(InputCategoriesRamp.EVENTS.RAMP_SELECTED, function () {
-      stackLayoutModel.goToStep(STEPS.RANGE_LIST_VIEW);
-    }, this);
-
-    view.on(InputCategoriesRamp.EVENTS.CUSTOM_COLOR, function () {
-      stackLayoutModel.goToStep(STEPS.RANGE_LIST_VIEW);
-    }, this);
+    // view.on(InputCategoriesRamp.EVENTS.RAMP_SELECTED, function () {
+    //   stackLayoutModel.goToStep(STEPS.CATEGORIES_LIST_VIEW);
+    // }, this);
+    //
+    // view.on(InputCategoriesRamp.EVENTS.CUSTOM_COLOR, function () {
+    //   stackLayoutModel.goToStep(STEPS.CATEGORIES_LIST_VIEW);
+    // }, this);
 
     return view;
   },
@@ -306,7 +314,7 @@ module.exports = CoreView.extend({
     });
 
     view.bind('back', function (value) {
-      stackLayoutModel.goToStep(STEPS.RANGE_LIST_VIEW);
+      stackLayoutModel.goToStep(STEPS.CATEGORIES_LIST_VIEW);
     }, this);
 
     view.bind('goToColorPicker', function () {
@@ -330,6 +338,11 @@ module.exports = CoreView.extend({
     this.model.set('images', images);
   },
 
+  _onClickQuantification: function (e) {
+    this.killEvent(e);
+    this.trigger('selectQuantification', this);
+  },
+
   _onClickBack: function (e) {
     this.killEvent(e);
     this.trigger('back', this);
@@ -337,3 +350,10 @@ module.exports = CoreView.extend({
 }, {
   MAX_VALUES: MAX_VALUES
 });
+
+function _quote (c) {
+  if (c && c !== true) {
+    return '"' + c.toString().replace(/"/g, '\\"').replace(/\n/g, '\\n') + '"';
+  }
+  return c;
+}
