+++ lib/assets/core/javascripts/cartodb3/components/form-components/editors/fill/input-color/input-quantitative-ramps/main-view.js
