+++ lib/assets/core/javascripts/cartodb3/editor/style/style-converter.js
@@ -1,7 +1,7 @@
 var _ = require('underscore');
 var camshaftReference = require('../../data/camshaft-reference');
 var Utils = require('../../helpers/utils');
-var InputColorCategories = require('../../components/form-components/editors/fill/input-color/input-categories/input-color-categories.js');
+var InputColorCategories = require('../../components/form-components/editors/fill/input-color/input-qualitative-ramps/main-view.js');
 
 var CONFIG = {
   GENERIC_STYLE: camshaftReference.getDefaultCartoCSSForType(),
@@ -53,7 +53,7 @@ function makeColorRamp (props, isTorqueCategory) {
             // Maps api converts null or empty value in emtpy string
             // and in the fill component we label them with a locale
             // so we use the same locale to generate the cartocss
-            return val === '' ? _t('form-components.editors.fill.input-categories.null') : val;
+            return val === '' ? _t('form-components.editors.fill.input-qualitative-ramps.null') : val;
           }),
           function (val) {
             return !_.isUndefined(val);
