+++ lib/assets/core/test/spec/cartodb3/components/form-components/editors/fill/input-color/input-categories/input-color-categories-list-view.spec.js
@@ -1,7 +1,7 @@
 var Backbone = require('backbone');
-var InputColorCategoriesList = require('../../../../../../../../../javascripts/cartodb3/components/form-components/editors/fill/input-color/input-categories/input-color-categories-list-view/input-color-categories-list-view');
+var InputColorCategoriesList = require('../../../../../../../../../javascripts/cartodb3/components/form-components/editors/fill/input-color/input-qualitative-ramps/input-categories-list/input-categories-list');
 
-describe('components/form-components/editors/fill/input-color/input-categories/input-color-categories-list-view', function () {
+describe('components/form-components/editors/fill/input-color/input-qualitative-ramps/input-categories-list', function () {
   beforeEach(function () {
     this.model = new Backbone.Model({
       domain: ['foo', 'bar', 'baz'],
