+++ lib/assets/core/test/spec/cartodb3/components/form-components/editors/fill/input-color/input-categories/input-color-categories-list-view/categories-list/categories-list-item-view.spec.js
@@ -1,8 +1,8 @@
 var CustomListItemModel = require('../../../../../../../../../../../javascripts/cartodb3/components/custom-list/custom-list-item-model');
-var CategoriesListItemView = require('../../../../../../../../../../../javascripts/cartodb3/components/form-components/editors/fill/input-color/input-categories/input-color-categories-list-view/categories-list/categories-list-item-view');
-var inputColorCategoriesListItemTemplate = require('../../../../../../../../../../../javascripts/cartodb3/components/form-components/editors/fill/input-color/input-categories/input-color-categories-list-view/input-color-categories-list-item.tpl');
+var CategoriesListItemView = require('../../../../../../../../../../../javascripts/cartodb3/components/form-components/editors/fill/input-color/input-qualitative-ramps/input-categories-list/categories-list/categories-list-item-view');
+var inputColorCategoriesListItemTemplate = require('../../../../../../../../../../../javascripts/cartodb3/components/form-components/editors/fill/input-color/input-qualitative-ramps/input-categories-list/input-color-categories-list-item.tpl');
 
-describe('components/form-components/editors/fill/input-color/input-categories/categories-list/categories-list-item-view', function () {
+describe('components/form-components/editors/fill/input-color/input-qualitative-ramps/categories-list/categories-list-item-view', function () {
   beforeEach(function () {
     this.model = new CustomListItemModel({
       label: 'hello',
