+++ lib/assets/core/test/spec/cartodb3/components/form-components/editors/fill/input-color/input-categories/input-color-categories.spec.js
@@ -1,12 +1,12 @@
 var Backbone = require('backbone');
-var InputColorCategories = require('../../../../../../../../../javascripts/cartodb3/components/form-components/editors/fill/input-color/input-categories/input-color-categories');
+var InputColorCategories = require('../../../../../../../../../javascripts/cartodb3/components/form-components/editors/fill/input-color/input-qualitative-ramps/main-view');
 var ConfigModel = require('../../../../../../../../../javascripts/cartodb3/data/config-model');
 var UserModel = require('../../../../../../../../../javascripts/cartodb3/data/user-model');
 var FactoryModals = require('../../../../../../factories/modals');
 var rampList = require('cartocolor');
 var cdb = require('cartodb.js');
 
-describe('components/form-components/editors/fill/input-color/input-categories/input-color-categories', function () {
+describe('components/form-components/editors/fill/input-color/input-qualitative-ramps/main-view', function () {
   beforeEach(function () {
     this.configModel = new ConfigModel({
       base_url: '/u/pepe',
