+++ lib/assets/core/test/spec/cartodb3/components/form-components/editors/fill/input-color/input-ramps/input-color-ramps.spec.js
@@ -1,7 +1,7 @@
 var Backbone = require('backbone');
-var InputColorRamps = require('../../../../../../../../../javascripts/cartodb3/components/form-components/editors/fill/input-color/input-quantitative-ramps/input-color-ramps');
+var InputColorRamps = require('../../../../../../../../../javascripts/cartodb3/components/form-components/editors/fill/input-color/input-quantitative-ramps/main-view');
 
-describe('components/form-components/editors/fill/input-color/input-quantitative-ramps/input-color-ramps', function () {
+describe('components/form-components/editors/fill/input-color/input-quantitative-ramps/main-view', function () {
   describe('on model with a range', function () {
     beforeEach(function () {
       this.model = new Backbone.Model({
